feat: finished second v/f the app, improved styles, project architecture
